requirements.txt
duckduckgo_search
requests
beautifulsoup4
readability-lxml
lxml
dateparser
pandas
tldextract
